+++ src/js/Icon.js
@@ -41,7 +41,7 @@ ctx.fill();
  *
  * @param text: text to draw (String or Number)
  */
-export function drawIcon(text) {
+export function draw(text) {
 	text = text.toString();
 	// Clear the inner part of the icon, without ever redrawing the border
 	ctx.clearRect(borderWidth, borderWidth, innerSize, innerSize);